Refactored uploadService to put modifying methods in a separate class, refactored some problematic methods with too much cognitive complexity
